Fixed duplication bug resulting in orphaned ids
The issue was with the `processedComponentsThisFrame` being cleared, well, each frame. Because the `ObjectChangeEventStream` was occurring across multiple frames the `PersistentIdManager` was losing track of the previously processed components from the same duplication operation, resulting in orphaned ids.

To fix: changed `processedComponentsThisFrame` to `processedComponentsThisDomainCycle` and only cleared it in the `[InitializeOnLoadMethod] Initialize()` method.