Fixed #1
It appears that setting a serialized property of type .intValue to a value above the int.MaxValue, for instance the uint returned by `registry.GenerateUniqueId()` will default to zero rather than implicitly detect that the backing field is actually a uint.

To fix: changed the idProp.intValue to idProp.uintValue when setting the PersistentId serialized property.